simplify webpack config minify usage
